move axdata to structs
